test: address overlay toast title followups
